refactor(builder-api,events-service): gate the axum servers behind a server feature
Both crates are client dependencies (BuilderClient via hotshot-task-impls,
the events Error via hotshot-query-service-types), and the port made axum,
http-wire's server stack, and url unconditional for every consumer. Follow
http-wire's own split: the routers live behind an off-by-default server
feature, enabled only by the crates that actually serve them (builder,
hotshot-builder-refactored, hotshot-builder-shared, hotshot-testing). The
events error types stay unconditional; the server half of events.rs moves to
events/server.rs so the gate is one cfg.
