+++ hotshot-events-service/Cargo.toml
@@ -5,28 +5,34 @@ edition.workspace = true
 license = "MIT"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[features]
+# The axum routers; off by default so client-only consumers (e.g. error-type imports) do not
+# build the server stack.
+server = ["dep:axum", "dep:http-wire", "dep:url"]
+
 [dependencies]
 alloy = { workspace = true }
 async-broadcast = { workspace = true }
 async-lock = { workspace = true }
 async-trait = { workspace = true }
-axum = { workspace = true }
+axum = { workspace = true, optional = true }
 derive_more = { workspace = true }
 disco-types = { workspace = true }
 futures = { workspace = true }
 hotshot-types = { workspace = true }
 http-client = { workspace = true }
-http-wire = { workspace = true, features = ["server"] }
+http-wire = { workspace = true, features = ["server"], optional = true }
 rand = { workspace = true }
 serde = { workspace = true }
 snafu = { workspace = true }
 tokio = { workspace = true }
 tracing = { workspace = true }
 tracing-test = { workspace = true }
-url = { workspace = true }
+url = { workspace = true, optional = true }
 vbs = { workspace = true }
 
 [dev-dependencies]
+hotshot-events-service = { path = ".", features = ["server"] }
 hotshot-example-types = { workspace = true }
 test-utils = { workspace = true }
